Fix [https://trac.osgeo.org/mapguide/ticket/291].
Fix some issues with point labeling using the 'EightSurrounding' algorithm.

* the text labels were not positioned correctly around the symbol (due to a regression)
* the code did not correctly handle the case with Y-down (GD renderer)
* there was a memory leak - one SE_RenderPointStyle per label
* added checks on the geometry type and symbol


git-svn-id: http://svn.osgeo.org/mapguide/trunk@2475 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Common/Stylization/SE_LinearClip.h
@@ -135,13 +135,13 @@ template<class CLIP_INFO>
 {
     if ((dst_cw || dst_ccw) && src->cntr_count())
     {
-        _ASSERT(src->geom_type() == (int)FdoGeometryType_LineString ||
-                src->geom_type() == (int)FdoGeometryType_MultiLineString ||
-                src->geom_type() == (int)FdoGeometryType_Polygon ||
-                src->geom_type() == (int)FdoGeometryType_MultiPolygon);
+        _ASSERT(src->geom_type() == FdoGeometryType_LineString ||
+                src->geom_type() == FdoGeometryType_MultiLineString ||
+                src->geom_type() == FdoGeometryType_Polygon ||
+                src->geom_type() == FdoGeometryType_MultiPolygon);
 
-        if (src->geom_type() == (int)FdoGeometryType_LineString ||
-            src->geom_type() == (int)FdoGeometryType_MultiLineString)
+        if (src->geom_type() == FdoGeometryType_LineString ||
+            src->geom_type() == FdoGeometryType_MultiLineString)
         {
             /*
              * The ClipPolygon function calculates the cw and ccw portions sepearately,
